ConDec-446: Move macro classes to upper view package
